tests/wpt/tests/css/css-anchor-position/inset-area-computed.html

<!DOCTYPE html>
<title>CSS Anchor Positioning: inset-area getComputedStyle()</title>
<link rel="help" href="https://drafts.csswg.org/css-anchor-position/#inset-area">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
<script src="/css/support/inheritance-testcommon.js"></script>
<div id="container">
  <div id="target"></div>
</div>
<script>
  test_computed_value("inset-area", "none");
  test_computed_value("inset-area", "span-all");
  test_computed_value("inset-area", "x-start");
  test_computed_value("inset-area", "center");
  test_computed_value("inset-area", "span-all span-all", "span-all");
  test_computed_value("inset-area", "center center", "center");
  test_computed_value("inset-area", "top center", "center top");
  test_computed_value("inset-area", "span-bottom span-all", "span-bottom");

  assert_not_inherited("inset-area", "none", "span-all");
</script>
